export from icon module
