+++ packages/cdc-react-icons/src/components/ArrowDown.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const ArrowDown = ({ className }: IconProps) => {
   return (
