+++ packages/cdc-react-icons/src/components/ArrowLeft.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const ArrowLeft = ({ className }: IconProps) => {
   return (
