+++ packages/cdc-react-icons/src/components/ArrowUp.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const ArrowUp = ({ className }: IconProps) => {
   return (
