+++ packages/cdc-react-icons/src/components/BarGraph.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const BarGraph = ({ className }: IconProps) => {
   return (
