+++ packages/cdc-react-icons/src/components/Bolt.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Bolt = ({ className }: IconProps) => {
   return (
