+++ packages/cdc-react-icons/src/components/Building.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Building = ({ className }: IconProps) => {
   return (
