+++ packages/cdc-react-icons/src/components/Calendar.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Calendar = ({ className }: IconProps) => {
   return (
