+++ packages/cdc-react-icons/src/components/Check.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Check = ({ className }: IconProps) => {
   return (
