+++ packages/cdc-react-icons/src/components/ChevronDoubleRight.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const ChevronDoubleRight = ({ className }: IconProps) => {
   return (
