+++ packages/cdc-react-icons/src/components/ChevronUp.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const ChevronUp = ({ className }: IconProps) => {
   return (
