+++ packages/cdc-react-icons/src/components/Close.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Close = ({ className }: IconProps) => {
   return (
