+++ packages/cdc-react-icons/src/components/CloseSquare.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const CloseSquare = ({ className }: IconProps) => {
   return (
