+++ packages/cdc-react-icons/src/components/Cloud.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Cloud = ({ className }: IconProps) => {
   return (
