+++ packages/cdc-react-icons/src/components/CloudArrowDown.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const CloudArrowDown = ({ className }: IconProps) => {
   return (
