+++ packages/cdc-react-icons/src/components/CloudArrowUp.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const CloudArrowUp = ({ className }: IconProps) => {
   return (
