+++ packages/cdc-react-icons/src/components/Cube.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Cube = ({ className }: IconProps) => {
   return (
