+++ packages/cdc-react-icons/src/components/Dots.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Dots = ({ className }: IconProps) => {
   return (
