+++ packages/cdc-react-icons/src/components/Eye.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Eye = ({ className }: IconProps) => {
   return (
