+++ packages/cdc-react-icons/src/components/Filter.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Filter = ({ className }: IconProps) => {
   return (
