+++ packages/cdc-react-icons/src/components/Folder.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Folder = ({ className }: IconProps) => {
   return (
