+++ packages/cdc-react-icons/src/components/HandPointArcs.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const HandPointArcs = ({ className }: IconProps) => {
   return (
