+++ packages/cdc-react-icons/src/components/House.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const House = ({ className }: IconProps) => {
   return (
