+++ packages/cdc-react-icons/src/components/LockShield.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const LockShield = ({ className }: IconProps) => {
   return (
