+++ packages/cdc-react-icons/src/components/Logout.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Logout = ({ className }: IconProps) => {
   return (
