+++ packages/cdc-react-icons/src/components/Menu.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Menu = ({ className }: IconProps) => {
   return (
