+++ packages/cdc-react-icons/src/components/Meter.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Meter = ({ className }: IconProps) => {
   return (
