+++ packages/cdc-react-icons/src/components/Minus.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Minus = ({ className }: IconProps) => {
   return (
