+++ packages/cdc-react-icons/src/components/PageLayout.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const PageLayout = ({ className }: IconProps) => {
   return (
