+++ packages/cdc-react-icons/src/components/PaperLines.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const PaperLines = ({ className }: IconProps) => {
   return (
