+++ packages/cdc-react-icons/src/components/Process.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Process = ({ className }: IconProps) => {
   return (
