+++ packages/cdc-react-icons/src/components/Quality.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Quality = ({ className }: IconProps) => {
   return (
