+++ packages/cdc-react-icons/src/components/Search.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Search = ({ className }: IconProps) => {
   return (
