+++ packages/cdc-react-icons/src/components/Settings.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Settings = ({ className }: IconProps) => {
   return (
