+++ packages/cdc-react-icons/src/components/SpeechBubble.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const SpeechBubble = ({ className }: IconProps) => {
   return (
