+++ packages/cdc-react-icons/src/components/Support.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Support = ({ className }: IconProps) => {
   return (
