+++ packages/cdc-react-icons/src/components/Swatch.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Swatch = ({ className }: IconProps) => {
   return (
