+++ packages/cdc-react-icons/src/components/Target.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const Target = ({ className }: IconProps) => {
   return (
