+++ packages/cdc-react-icons/src/components/TrafficCone.tsx
@@ -1,4 +1,4 @@
-import { IconProps } from ".";
+import { IconProps } from "./Icons";
 
 export const TrafficCone = ({ className }: IconProps) => {
   return (
